src/backend/requirements.txt
Django
django-cors-headers
djangorestframework
drf-yasg
psycopg2-binary
drf-yasg
djangorestframework-simplejwt
djoser
django-templated-mail
django-jazzmin
gunicorn
whitenoise
django-extensions
django-agenda
pytest
tzdata
django-silk
Pillow
requests